Seperate orderbooks by proposal
